refactor: action handlers around shared selector resolution
