+++ internal/handlers/action_types.go
@@ -1,0 +1,27 @@
+package handlers
+
+import "github.com/pinchtab/pinchtab/internal/bridge"
+
+type actionsRequest struct {
+	TabID       string                 `json:"tabId"`
+	Owner       string                 `json:"owner"`
+	Actions     []bridge.ActionRequest `json:"actions"`
+	StopOnError bool                   `json:"stopOnError"`
+}
+
+type actionResult struct {
+	Index   int            `json:"index"`
+	Success bool           `json:"success"`
+	Result  map[string]any `json:"result,omitempty"`
+	Error   string         `json:"error,omitempty"`
+}
+
+func countSuccessful(results []actionResult) int {
+	count := 0
+	for _, r := range results {
+		if r.Success {
+			count++
+		}
+	}
+	return count
+}
